requirements.txt
chainlit
langgraph
langchain
langchain-community
langchain-google-community[google-calendar]
langchain-litellm
litellm
python-dotenv
pydantic
google-auth-oauthlib
google-api-python-client
